implement SAF GetProperties API
:Release Notes:
implemented GetProperties API
for storageaccess service

:Detailed Notes:

:Testing Performed:
Tested with Engg builds

:QA Notes:
N/A

:Issues Addressed:
[PLAT-131331] [SAF] Implementation for
              GetProperties API wrt.
              internal storage

Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.service.storageaccess/+/304330
Tested-by: RajeshGopu IV <rajeshgopu.iv@lge.com>

+++ src/providers/internal/InternalOperationHandler.cpp
@@ -17,6 +17,8 @@
 #include <bitset>
 #include <iostream>
 #include <filesystem>
+#include "SAFLog.h"
+
 namespace fs = std::filesystem;
 
 FolderContent::FolderContent(std::string absPath) : mPath(absPath)
@@ -76,7 +78,7 @@ uint32_t FolderContent::getFileSize(std::string filePath)
     return size;
 }
 
-FolderContents::FolderContents(std::string fullPath) : mFullPath(fullPath)
+FolderContents::FolderContents(std::string fullPath) : mFullPath(fullPath), mStatus(0)
 {
     init();
 }
@@ -91,31 +93,39 @@ void FolderContents::init()
             fs::directory_options::skip_permission_denied);
         for (const auto & entry : fs::recursive_directory_iterator(mFullPath, fs::directory_options(options)))
         {
-            FolderContent folderObj = FolderContent(entry.path());
+            std::shared_ptr<FolderContent> folderObj = std::shared_ptr<FolderContent>(new FolderContent(entry.path()));
             mContents.push_back(folderObj);
+            //LOG_DEBUG_SAF("%s: Path: %s, Total: %d", __FUNCTION__, entry.path().c_str(), mContents.size());
         }
     }
     catch(fs::filesystem_error& e) {
         mTotalCount = 0;
+        mStatus = -1;
     }
     mTotalCount = mContents.size();
 }
 
-InternalSpaceInfo::InternalSpaceInfo(std::string path) : mPath(path)
+InternalSpaceInfo::InternalSpaceInfo(std::string path) : mPath(path), mStatus(0)
 {
     init();
 }
 
 void InternalSpaceInfo::init()
 {
-    fs::space_info dev = fs::space(mPath);
-    mCapacity = dev.capacity;
-    mFreeSpace = dev.free;
-    mAvailSpace = dev.available;
+    try{
+        fs::space_info dev = fs::space(mPath);
+        mCapacity = dev.capacity;
+        mFreeSpace = dev.free;
+        mAvailSpace = dev.available;
 
-    fs::perms ps = fs::status(mPath).permissions();
-    mIsWritable = ((ps & fs::perms::owner_write) != fs::perms::none);
-    mIsDeletable = ((ps & fs::perms::group_write) != fs::perms::none);
+        fs::perms ps = fs::status(mPath).permissions();
+        mIsWritable = ((ps & fs::perms::owner_write) != fs::perms::none);
+        mIsDeletable = ((ps & fs::perms::group_write) != fs::perms::none);
+        mStatus = 100;
+    }
+    catch(fs::filesystem_error& e) {
+        mStatus = -1;
+    }
 }
 
 std::uint32_t InternalSpaceInfo::getCapacityMB()
@@ -143,8 +153,14 @@ bool InternalSpaceInfo::getIsDeletable()
     return mIsDeletable;
 }
 
-InternalCopy::InternalCopy(std::string src, std::string dest)
-    : mSrcPath(src), mDestPath(dest), mStatus(0)
+int32_t InternalSpaceInfo::getStatus()
+{
+    return mStatus;
+}
+
+
+InternalCopy::InternalCopy(std::string src, std::string dest, bool overwrite)
+    : mSrcPath(src), mDestPath(dest), mStatus(0), mOverwrite(overwrite)
 {
     init();
 }
@@ -156,10 +172,14 @@ void InternalCopy::init()
     std::async(std::launch::async, [this]()
         {
             try {
-                const auto copyOptions = fs::copy_options::update_existing
-                           | fs::copy_options::recursive
+                auto copyOptions = fs::copy_options::recursive
                            | fs::copy_options::skip_symlinks;
-                fs::copy(this->mSrcPath, this->mDestPath, copyOptions); 
+                if (this->mOverwrite)
+                    copyOptions |= fs::copy_options::overwrite_existing;
+                else
+                    copyOptions |= fs::copy_options::skip_existing;
+                fs::copy(this->mSrcPath, this->mDestPath, copyOptions);
+                this->mStatus = 100;
             }
             catch(fs::filesystem_error& e) {
                 this->mStatus = -1;
@@ -170,9 +190,13 @@ void InternalCopy::init()
 std::uint32_t InternalCopy::getStatus()
 {
     uint32_t size = FolderContent(mDestPath).getSize();
-    uint32_t change = (size > mSrcSize)?(size - mSrcSize):(size - mDestSize);
-    if (mStatus != -1 && mSrcSize)
+    uint32_t change = (mSrcSize - size);
+    if ((mStatus != -1) && (change > 0))
         mStatus = int(change / mSrcSize * 100);
+    else if ((mStatus != -1) && (change == 0))
+        mStatus = 100;
+    else
+        mStatus = -1;
     return mStatus;
 }
 
@@ -185,8 +209,13 @@ InternalRemove::InternalRemove(std::string path)
 void InternalRemove::init()
 {
     try {
-        fs::remove_all(mPath);
-        mStatus = 100;
+        if (fs::exists(mPath))
+        {
+            fs::remove_all(mPath);
+            mStatus = 100;
+        }
+        else
+            mStatus = -1;
     }
     catch(fs::filesystem_error& e) {
         mStatus = -1;
@@ -198,34 +227,23 @@ int32_t InternalRemove::getStatus()
     return mStatus;
 }
 
-InternalMove::InternalMove(std::string srcPath, std::string destPath)
-    : mSrcPath(srcPath), mDestPath(destPath), mStatus(0)
+InternalMove::InternalMove(std::string srcPath, std::string destPath, bool overwrite)
+    : mSrcPath(srcPath), mDestPath(destPath), mStatus(0), mOverwrite(overwrite)
 {
     init();
 }
 
 void InternalMove::init()
 {
-    std::async(std::launch::async, [this]()
-        {
-            try {
-                InternalCopy obj = InternalCopy(this->mSrcPath, this->mDestPath);
-                this->mStatus = obj.getStatus();
-                while (this->mStatus!= -1)
-                {
-                    if (this->mStatus >= 100)
-                    {
-                        InternalRemove rem = InternalRemove(this->mSrcPath);
-                        this->mStatus = rem.getStatus();
-                        break;
-                    }
-                    this->mStatus = obj.getStatus();
-                }
-            }
-            catch(fs::filesystem_error& e) {
-                this->mStatus = -1;
-            }
-        });
+    InternalRename obj = InternalRename(mSrcPath, mDestPath);
+    mStatus = obj.getStatus();
+    if ((mStatus == -1) && mOverwrite)
+    {
+        InternalRemove remObj = InternalRemove(mDestPath);
+        mStatus = remObj.getStatus();
+        InternalRename retryObj = InternalRename(mSrcPath, mDestPath);
+        mStatus = retryObj.getStatus();
+    }
 }
 
 int32_t InternalMove::getStatus()
@@ -233,6 +251,27 @@ int32_t InternalMove::getStatus()
     return mStatus;
 }
 
+InternalRename::InternalRename(std::string oldAbsPath, std::string newAbsPath)
+    : mOldAbsPath(oldAbsPath), mNewAbsPath(newAbsPath), mStatus(0)
+{
+    init();
+}
+
+void InternalRename::init ()
+{
+    try {
+        fs::rename(mOldAbsPath, mNewAbsPath);
+        mStatus = 100;
+    }
+    catch(fs::filesystem_error& e) {
+        mStatus = -1;
+    }
+}
+
+int32_t InternalRename::getStatus()
+{
+    return mStatus;
+}
 
 InternalOperationHandler::InternalOperationHandler()
 {
@@ -244,50 +283,41 @@ InternalOperationHandler& InternalOperationHandler::getInstance()
     return obj;
 }
 
-std::shared_ptr<FolderContents> InternalOperationHandler::getListFolderContents(std::string path)
+std::unique_ptr<FolderContents> InternalOperationHandler::getListFolderContents(std::string path)
 {
-    std::shared_ptr<FolderContents> obj = std::make_shared<FolderContents>(
-path);
-    dumpContents();
-    return obj;
+    std::unique_ptr<FolderContents> obj = std::unique_ptr<FolderContents>( new FolderContents(path));
+    return std::move(obj);
 }
 
-void InternalOperationHandler::dumpContents()
+std::unique_ptr<InternalSpaceInfo> InternalOperationHandler::getProperties(std::string path)
 {
-    std::cout << "FullPath: " << mFolderContentObj->getPath() << ", Total: " << 
-mFolderContentObj->getTotalCount() << std::endl;
-    for (auto content : mFolderContentObj->getContents())
-    {
-        std::cout << "FileName: " << content.getName() << ", Path: " << 
-content.getPath()
-            << ", Type: " << content.getType() << ", Size: " << content.
-getSize() << std::endl;
-    }
+    if (path.empty())   path = "/tmp";
+    std::unique_ptr<InternalSpaceInfo> obj = std::unique_ptr<InternalSpaceInfo>( new InternalSpaceInfo(path));
+    return std::move(obj);
 }
 
-std::shared_ptr<InternalSpaceInfo> InternalOperationHandler::getProperties(std::string path)
+std::unique_ptr<InternalCopy> InternalOperationHandler::copy(std::string srcPath,
+    std::string destPath, bool overwrite)
 {
-    if (path.empty())
-        path = "/tmp";
-    mSpaceInfoObj = std::make_shared<InternalSpaceInfo>(path);
-    return mSpaceInfoObj;
+    std::unique_ptr<InternalCopy> obj = std::unique_ptr<InternalCopy>(new InternalCopy(srcPath, destPath, overwrite));
+    return std::move(obj);
 }
 
-std::shared_ptr<InternalCopy> InternalOperationHandler::copy(std::string srcPath, std::string destPath)
+std::unique_ptr<InternalRemove> InternalOperationHandler::remove(std::string path)
 {
-    mCopyObj = std::make_shared<InternalCopy>(srcPath, destPath);
-    return mCopyObj;
+    std::unique_ptr<InternalRemove> obj = std::unique_ptr<InternalRemove>(new InternalRemove(path));
+    return std::move(obj);
 }
 
-std::shared_ptr<InternalRemove> InternalOperationHandler::remove(std::string path)
+std::unique_ptr<InternalMove> InternalOperationHandler::move(std::string srcPath, std::string destPath, bool overwrite)
 {
-    mRemoveObj = std::make_shared<InternalRemove>(path);
-    return mRemoveObj;
+    std::unique_ptr<InternalMove> obj = std::unique_ptr<InternalMove>(new InternalMove(srcPath, destPath, overwrite));
+    return std::move(obj);
 }
 
-std::shared_ptr<InternalMove> InternalOperationHandler::move(std::string srcPath, std::string destPath)
+std::unique_ptr<InternalRename> InternalOperationHandler::rename(std::string srcPath, std::string destPath)
 {
-    mMoveObj = std::make_shared<InternalMove>(srcPath, destPath);
-    return mMoveObj;
+    std::unique_ptr<InternalRename> obj = std::unique_ptr<InternalRename>(new InternalRename(srcPath, destPath));
+    return std::move(obj);
 }
